requirements.txt
--extra-index-url https://download.pytorch.org/whl/cu128

# Existing packages (unchanged)
torch>=2.2.0
torchvision>=0.17.0
transformers>=4.40.0
accelerate>=0.28.0
diffusers>=0.27.0
peft>=0.10.0
datasets>=2.18.0
safetensors>=0.4.3
pyyaml>=6.0
omegaconf>=2.3.0
jsonschema>=4.21.0
einops>=0.7.0
pillow>=10.3.0
numpy>=1.26.0
tqdm>=4.66.0
rich>=13.7.0
psutil>=5.9.8
bitsandbytes>=0.43.0
triton>=2.2.0
flash-attn>=2.5.0; platform_system=="Linux"
xformers>=0.0.25; platform_system=="Linux"
wandb>=0.16.6
tensorboard>=2.16.0
torchmetrics>=1.3.2
scipy>=1.12.0

# === Added for Qwen3-VL video + image support ===
qwen-vl-utils[decord]>=0.0.14